+++ changelog.md
@@ -1,1 +1,9 @@
 - 2026-05-13: I made the base models for this project and set up a git project
+- 2026-05-14: We made the connection between mongodb and python
+- 2026-05-16: saved auth details to mongodb
+- 2026-05-18 made functions that are concerned with security of user login and signup. i import database variable from mongo_client.py instead of mongo_client variable. 
+
+Tested the security.py file with the unittest package
+
+added user data to be returned by sign-in route
+started a storage.py route that can store session data, and I tested it.. storage was meant for a new branch, though